backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
google-cloud-aiplatform==1.50.0
google-cloud-discoveryengine>=0.11.2  # For Vertex AI Search
google-auth==2.27.0
python-multipart==0.0.9               # For file uploads
pydantic-settings==2.1.0              # Pro configuration
firebase-admin==6.4.0                 # Auth validation
httpx==0.26.0                         # Async HTTP client
